move imagine components to all components
